+++ packages/ckb-sdk-wallet/CHANGELOG.md
@@ -8,6 +8,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
+* **feat:** use bigint to calculate capacity([b4cae6](https://github.com/nervosnetwork/ckb-sdk-js/commit/b4cae6))
 * **types:** add field of validSince in cell input of a transaction ([54770fc](https://github.com/nervosnetwork/ckb-sdk-js/commit/54770fc))
 * **types:** update ckb-types ([e6af3b5](https://github.com/nervosnetwork/ckb-sdk-js/commit/e6af3b5))